Go SDK auto pagination tweaks (#3697)
* api: remove deleted items from list response

When automatically paginating over the responses from
our list endpoints, there is a chance that an item could
be included both in the list of items and in the list of
deleted IDs (if it was deleted while paginating). Perform
a post-processing step to remove these items both
from the Items slice and the RemovedIds slice.

* api: update previously seen items in-place

When automatically paginating over the responses from
our list endpoints, there is a chance that an item
could be included twice (if it is updated during
pagination). If an item is seen twice, update it
in-place instead.

* fix PR issues

* api: remove page_size option

This is no longer necessary as we're always performing
full pagination in the SDK.

* api: use 1.19 compatible slices package

* api: add slices package dependency