Add credential libraries pagination (#3721)
* pagination: use size of items when possible

When we have a request without a request token and all
results fit within the page size, we can set the estimated
item count directly from the size of the items, since it
contains all visible items.

* pagination: assert we have enough after filtering

The previous code had a bug whereby it assumed that as
soon as there is at least one more element available from
the list items function, we're not at the end. This is incorrect,
as that element may be subject to the filter.

The new code asserts that we have at least one more element
available after the filtering has been applied. This allows us
to correctly determine when a listing is complete and not.

This also allowed us to significantly simplify the fillPage function.

* proto: tidy up some comments

* credentiallibraries: add pagination