all: refactor credential library listing (#3740)
The new structure separates the methods
that operate on a generic credential library entity
out of the vault package and into the
credential package. This isn't strictly
necessary for credential libraries since
they all live in the vault repo at the moment, but
this structure is necessary for credential stores
and it makes more sense from a structure point
to do it for both resources.